refactor: organize given for symbol generation
