+++ src/main/scala/satify/dsl/DSL.scala
@@ -1,7 +1,12 @@
 package satify.dsl
 
+import satify.model.expression.SymbolGeneration.{SymbolGenerator, encodingVarPrefix}
+
 object DSL:
-  export Conversion.given
+  export Conversions.given
   export Operators.*
   export Numbers.*
   export Encodings.*
+
+  given SymbolGenerator with
+    def prefix: String = encodingVarPrefix
