+++ src/main/scala/satify/model/problems/NQueens.scala
@@ -5,15 +5,16 @@ import satify.model.dpll.{PartialModel, Variable}
 import satify.model.expression.Encodings.{atLeastOne, atMostOne}
 import satify.model.expression.Expression
 import satify.model.expression.Expression.*
+import satify.model.expression.SymbolGeneration.{SymbolGenerator, encodingVarPrefix}
 
 import scala.annotation.tailrec
 import scala.swing.{Component, FlowPanel}
 
 case class NQueens(n: Int) extends Problem:
 
-  given IncrementalSymbolGenerator with
-    override def prefix: String = "ENC"
-
+  given SymbolGenerator with
+    def prefix: String = encodingVarPrefix
+    
   private val variables: Seq[Seq[Symbol]] =
     if n < 0 then throw new IllegalArgumentException("n must be positive")
     else
