+++ src/main/scala/satify/model/problems/NurseScheduling.scala
@@ -1,14 +1,15 @@
 package satify.model.problems
 import satify.model.expression.Expression
 import satify.model.expression.Expression.*
+import satify.model.expression.SymbolGeneration.{SymbolGenerator, encodingVarPrefix}
 
 import scala.swing.{Component, FlowPanel}
 
 case class NurseScheduling(nurses: Int, days: Int, shifts: Int) extends Problem:
 
-  given IncrementalSymbolGenerator with
-    override def prefix: String = "ENC"
-    
+  given SymbolGenerator with
+    def prefix: String = encodingVarPrefix
+
   private val variables: Seq[Seq[Seq[Symbol]]] =
     for n <- 0 until nurses
     yield for d <- 0 until days
